rework video mode list composition (bsc#1096971)
The major change here is to accept any modes with >= 16 color bits and to
prefer the one with the highest number of color bits.

The old code looked only for 16 bit modes.

This gives a wider range of video modes to pick from and reduces the chance
to end up empty-handed.

For this the mode array was extended with a new field '.vm_bits' to track
the color size.

'vmsortindex' was replaced by 'vmcmp' to take color size into account. The
rest of the changes deal basically with adjusting to the new mode array size.

I've left the two debug functions in - it will not hurt in code size as they
get optimized out when unused. But it's convenient in case they are needed
again.