limit the 'autodown' feature to first key press
If config.autodown is set, the main menu would automatically move one menu
entry down the first time an F-key is used.

This was originally intended to keep people from selecting the 'local boot'
option (usually first entry) accidentally.

But this has unwanted side effects if you go via 'more...' to a second level
menu: it would do the same there.

To avoid this, limit this feature to the first key press.